refactor : coredata 저장 로직을 직접객체를 만드는 형태에서 NSManagedObject를 만드는 방식으로 변경
- 측정소요시간 저장 방식 수정
- MeasureViewController를 만들 때 DataListView에서 가지고 있던 service를 주입
- cell의 소요시간 표시를 소숫점 첫째자리까지 표현
- coreData type의 값을 Double로 수정

+++ SensorData+CoreDataProperties.swift
@@ -19,7 +19,7 @@ extension SensorData {
     @NSManaged public var xValue: [Double]
     @NSManaged public var yValue: [Double]
     @NSManaged public var zValue: [Double]
-    @NSManaged public var type: Int16
+    @NSManaged public var type: Double
     @NSManaged public var runTime: Double
     @NSManaged public var date: Date
